backend/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
django-cors-headers==3.5.0
psycopg2==2.8.6
igdb-api-v4==0.0.3